Removing Interfaces from External Files
