--- onboard-jmuniz/constants/interfaces/login-text-input-props.ts
@@ -1,4 +1,0 @@
-export interface LoginTextInputProps {
-  label: string;
-  conditions: { pattern: RegExp; message: string }[];
-}
