--- onboard-jmuniz/constants/interfaces/login-user-response-props.ts
@@ -1,5 +1,0 @@
-export interface LoginUserResponseProps {
-  login: {
-    token: string;
-  };
-}
